Move WAC connect view to the WhatsApp cloud channel type package
