+++ temba/channels/types/whatsapp_cloud/views.py
@@ -32,10 +32,10 @@ class ClaimView(ClaimViewMixin, SmartFormView):
     form_class = Form
 
     def pre_process(self, request, *args, **kwargs):
-        oauth_user_token = self.request.session.get(Channel.CONFIG_WHATSAPP_CLOUD_USER_TOKEN, None)
+        oauth_user_token = self.request.session.get(self.channel_type.SESSION_WHATSAPP_CLOUD_USER_TOKEN, None)
         if not oauth_user_token:
             self.remove_token_credentials_from_session()
-            return HttpResponseRedirect(reverse("orgs.org_whatsapp_cloud_connect"))
+            return HttpResponseRedirect(reverse("channels.types.whatsapp_cloud.connect"))
 
         app_id = settings.FACEBOOK_APPLICATION_ID
         app_secret = settings.FACEBOOK_APPLICATION_SECRET
@@ -46,13 +46,13 @@ class ClaimView(ClaimViewMixin, SmartFormView):
         response = requests.get(url, params=params)
         if response.status_code != 200:  # pragma: no cover
             self.remove_token_credentials_from_session()
-            return HttpResponseRedirect(reverse("orgs.org_whatsapp_cloud_connect"))
+            return HttpResponseRedirect(reverse("channels.types.whatsapp_cloud.connect"))
 
         response_json = response.json()
         for perm in ["business_management", "whatsapp_business_management", "whatsapp_business_messaging"]:
             if perm not in response_json["data"]["scopes"]:
                 self.remove_token_credentials_from_session()
-                return HttpResponseRedirect(reverse("orgs.org_whatsapp_cloud_connect"))
+                return HttpResponseRedirect(reverse("channels.types.whatsapp_cloud.connect"))
 
         return super().pre_process(request, *args, **kwargs)
 
@@ -62,7 +62,7 @@ class ClaimView(ClaimViewMixin, SmartFormView):
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
 
-        oauth_user_token = self.request.session.get(Channel.CONFIG_WHATSAPP_CLOUD_USER_TOKEN, None)
+        oauth_user_token = self.request.session.get(self.channel_type.SESSION_WHATSAPP_CLOUD_USER_TOKEN, None)
         app_id = settings.FACEBOOK_APPLICATION_ID
         app_secret = settings.FACEBOOK_APPLICATION_SECRET
 
@@ -126,7 +126,7 @@ class ClaimView(ClaimViewMixin, SmartFormView):
 
         context["claim_url"] = reverse("channels.types.whatsapp_cloud.claim")
         context["clear_session_token_url"] = reverse("channels.types.whatsapp_cloud.clear_session_token")
-        context["connect_whatsapp_url"] = reverse("orgs.org_whatsapp_cloud_connect")
+        context["connect_whatsapp_url"] = reverse("channels.types.whatsapp_cloud.connect")
         context["facebook_app_id"] = settings.FACEBOOK_APPLICATION_ID
 
         claim_error = None
@@ -209,16 +209,16 @@ class ClaimView(ClaimViewMixin, SmartFormView):
         return super().form_valid(form)
 
     def remove_token_credentials_from_session(self):
-        if Channel.CONFIG_WHATSAPP_CLOUD_USER_TOKEN in self.request.session:
-            del self.request.session[Channel.CONFIG_WHATSAPP_CLOUD_USER_TOKEN]
+        if self.channel_type.SESSION_WHATSAPP_CLOUD_USER_TOKEN in self.request.session:
+            del self.request.session[self.channel_type.SESSION_WHATSAPP_CLOUD_USER_TOKEN]
 
 
 class ClearSessionToken(ChannelTypeMixin, OrgPermsMixin, SmartTemplateView):
     permission = "channels.channel_claim"
 
     def pre_process(self, request, *args, **kwargs):
-        if Channel.CONFIG_WHATSAPP_CLOUD_USER_TOKEN in self.request.session:
-            del self.request.session[Channel.CONFIG_WHATSAPP_CLOUD_USER_TOKEN]
+        if self.channel_type.SESSION_WHATSAPP_CLOUD_USER_TOKEN in self.request.session:
+            del self.request.session[self.channel_type.SESSION_WHATSAPP_CLOUD_USER_TOKEN]
 
     def render_to_response(self, context, **response_kwargs):
         return JsonResponse({})
@@ -346,3 +346,77 @@ class VerifyCode(ChannelTypeMixin, ModalMixin, ContentMenuMixin, OrgObjPermsMixi
                 _("Unable to register phone %s with ID %s from WABA with ID %s")
                 % (wa_number, channel.address, waba_id)
             )
+
+
+class Connect(ChannelTypeMixin, OrgPermsMixin, SmartFormView):
+    class WhatsappCloudConnectForm(forms.Form):
+        user_access_token = forms.CharField(min_length=32, required=True)
+
+        def clean(self):
+            try:
+                auth_token = self.cleaned_data.get("user_access_token", None)
+
+                app_id = settings.FACEBOOK_APPLICATION_ID
+                app_secret = settings.FACEBOOK_APPLICATION_SECRET
+
+                url = "https://graph.facebook.com/v13.0/debug_token"
+                params = {"access_token": f"{app_id}|{app_secret}", "input_token": auth_token}
+
+                response = requests.get(url, params=params)
+                if response.status_code != 200:  # pragma: no cover
+                    raise Exception("Failed to debug user token")
+
+                response_json = response.json()
+
+                for perm in ["business_management", "whatsapp_business_management", "whatsapp_business_messaging"]:
+                    if perm not in response_json.get("data", dict()).get("scopes", []):
+                        raise Exception(
+                            'Missing permission, we need all the following permissions "business_management", "whatsapp_business_management", "whatsapp_business_messaging"'
+                        )
+            except Exception:
+                raise forms.ValidationError(
+                    _("Sorry account could not be connected. Please try again"), code="invalid"
+                )
+
+            return self.cleaned_data
+
+    permission = "channels.types.whatsapp_cloud.connect"
+    form_class = WhatsappCloudConnectForm
+    success_url = "@channels.types.whatsapp_cloud.claim"
+    field_config = dict(api_key=dict(label=""), api_secret=dict(label=""))
+    submit_button_name = "Save"
+    success_message = "WhatsApp Account successfully connected."
+    template_name = "channels/types/whatsapp_cloud/connect.html"
+    menu_path = "/settings/workspace"
+    title = "Connect WhatsApp"
+
+    def has_org_perm(self, permission):
+        if self.org:
+            return self.get_user().is_beta  # only beta users are allowed
+        return False
+
+    def pre_process(self, request, *args, **kwargs):
+        session_token = self.request.session.get(self.channel_type.SESSION_WHATSAPP_CLOUD_USER_TOKEN, None)
+        if session_token:
+            return HttpResponseRedirect(self.get_success_url())
+
+        return super().pre_process(request, *args, **kwargs)
+
+    def form_valid(self, form):
+        auth_token = form.cleaned_data["user_access_token"]
+
+        # add the credentials to the session
+        self.request.session[self.channel_type.SESSION_WHATSAPP_CLOUD_USER_TOKEN] = auth_token
+        return HttpResponseRedirect(self.get_success_url())
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["connect_url"] = reverse("channels.types.whatsapp_cloud.connect")
+        context["facebook_app_id"] = settings.FACEBOOK_APPLICATION_ID
+
+        claim_error = None
+        if context["form"].errors:
+            claim_error = context["form"].errors.get("__all__", [""])[0]
+        context["claim_error"] = claim_error
+
+        return context
