+++ temba/settings_common.py
@@ -425,7 +425,6 @@ PERMISSIONS = {
         "twilio_account",
         "twilio_connect",
         "two_factor",
-        "whatsapp_cloud_connect",
         "workspace",
     ),
     "request_logs.httplog": ("webhooks", "classifier", "ticketer"),
@@ -441,7 +440,7 @@ PERMISSIONS = {
 GROUP_PERMISSIONS = {
     "Service Users": ("flows.flow_assets", "msgs.msg_create"),  # internal Temba services have limited permissions
     "Alpha": (),
-    "Beta": ("orgs.org_whatsapp_cloud_connect",),
+    "Beta": (),
     "Dashboard": ("orgs.org_dashboard",),
     "Surveyors": (
         "contacts.contact_api",
